consolidate UpdateCallback in ./renderpanel
